+++ hata/discord/voice/encryption_adapters/__init__.py
@@ -1,6 +1,7 @@
 from .aead_aes256_gcm_rtpsize import *
 from .aead_xchacha20_poly1305_rtpsize import *
 from .base import *
+from .nacl import *
 from .xsalsa20_poly1305 import *
 
 
@@ -8,6 +9,7 @@ __all__ = (
     *aead_aes256_gcm_rtpsize.__all__,
     *aead_xchacha20_poly1305_rtpsize.__all__,
     *base.__all__,
+    *nacl.__all__,
     *xsalsa20_poly1305.__all__,
 )