+++ hata/ext/slash/command/command_base_custom_id/tests/test__split_and_check_satisfaction.py
@@ -16,7 +16,7 @@ def _create_parameter(name, *, annotation = ..., default = ...):
     Parameters
     ----------
     name : `str`
-        The name of teh parameter.
+        The name of the parameter.
     
     annotation : `object`, Optional (Keyword only)
         Annotation of the parameter.